Update community
We can't use 'framework' without explaining it. I'm not sure what this is. I think an introduction (what this document does) rather than 'context' (justification for the document) would help.

This is very jargon-heavy: It is intended to provide community owners and members with a common vocabulary and set of approaches, aimed at helping communities to function more effectively and efficiently, as well as providing support to enable ongoing development and capability-building.
What does it mean?

Everything under the H2 ## Community typology needs rewriting. Types of communities are described and then some of them are developed under their own H2s. We need to be consistent. 

It'll be worth reading the style guide on bullet points and steps.https://www.gov.uk/design-principles/style-guide/style-points#style-bullets